requirements.txt
# Core
flask>=3.0.0
python-dotenv>=1.0.0

# Database
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0

# Forms & Validation
flask-wtf>=1.2.0
wtforms>=3.1.0

# Authentication & Security
flask-login>=0.6.3
bcrypt>=4.0.0
argon2-cffi>=23.0.0
flask-limiter>=3.5.0
flask-cors>=4.0.0
markupsafe>=2.1.3

# Excel/PDF Generation
openpyxl>=3.1.0
reportlab>=4.0.0

# OCR / Google Cloud Vision
google-cloud-vision>=3.5.0

# Development
pytest>=7.4.0
ruff>=0.1.0
